salvataggio stato allenamenti, possibilità di ricaricare il timer

+++ app/src/main/java/com/example/progetto_rosso_iacopo/data/model/WorkoutRoutine.kt
@@ -6,7 +6,7 @@ data class Exercise(
     val name: String = "",
     val reps: Int = 0,
     val sets: Int = 0,
-    val restTimeSeconds: Int = 0,
+    val restTimeSeconds: Long = 0,
     val description: String = "",
 ) {
     companion object {
@@ -15,7 +15,7 @@ data class Exercise(
             name: String,
             reps: Int?,
             sets: Int?,
-            restTimeSeconds: Int?,
+            restTimeSeconds: Long?,
             description: String?
         ): Exercise {
             return Exercise(
